+++ tests/TestImport.hs
@@ -2,10 +2,12 @@ module TestImport
   ( fromJust,
     module Control.Monad.Fail,
     module Test.Hspec,
+    module Test.Hspec.QuickCheck,
     module ClassyPrelude,
     module Data.Aeson,
     module Data.Aeson.TH,
     cs,
+    module Test.QuickCheck,
   )
 where
 
@@ -16,3 +18,6 @@ import Data.Aeson.TH (deriveJSON)
 import Data.Maybe (fromJust)
 import Data.String.Conversions (cs)
 import Test.Hspec
+import Test.Hspec.QuickCheck
+import Test.QuickCheck
+import Test.QuickCheck.Instances ()