+++ tests/Web/Slack/Experimental/Events/TypesSpec.hs
@@ -24,4 +24,6 @@ spec = describe "Types for Slack events" do
         , "channel_left"
         , "share_without_message"
         , "circleci"
+        , -- https://slack.com/help/articles/206819278-Send-emails-to-Slack
+          "email_message"
         ]