+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.6.2.0 (2023-10-24)
+
+* [#127](https://github.com/MercuryTechnologies/slack-web/pull/127)
+  Add helper for the divider block type.
+
+# 1.6.1.0 (2022-12-16)
+
+* [#124](https://github.com/MercuryTechnologies/slack-web/pull/124)
+  Parse [Slack incoming emails](https://slack.com/help/articles/206819278-Send-emails-to-Slack)
+  without throwing an error.
+
 # 1.6.0.0 (2022-12-14)
 
 ## Breaking changes